Fix type equals issue in test

+++ core/genesis_test.go
@@ -190,7 +190,7 @@ func TestStatefulPrecompilesConfigure(t *testing.T) {
 			},
 			assertState: func(t *testing.T, sdb *state.StateDB) {
 				assert.Equal(t, precompile.AllowListAdmin, precompile.GetAllowListStatus(sdb, addr), "unexpected allow list status for modified address")
-				assert.Equal(t, 1, sdb.GetNonce(precompile.AllowListAddress))
+				assert.Equal(t, uint64(1), sdb.GetNonce(precompile.AllowListAddress))
 			},
 		},
 	} {
